Cleanup extra component files
