+++ packages/site/src/pages/card/[designId].astro
@@ -7,7 +7,7 @@ import { routes, routeNames, LEGACY_CARD_ID, ASSETS } from '@site/constants';
 import { PageTitle } from '@site/components/text';
 import { Heading } from '@site/components/text';
 import DesignInstancesCardList from '@site/components/cards/DesignInstancesCardList';
-import { parseFiltersFromSearchParams } from '@site/components/cards/CardListFilter';
+import { parseFiltersFromSearchParams } from '@site/components/cards/CardList';
 
 const designId = Astro.params.designId;
