+++ packages/site/src/pages/card/index.astro
@@ -4,7 +4,7 @@ import AllDesignsCardList from '@site/components/cards/AllDesignsCardList';
 import { getAllCardDesigns } from '@core/lib/design';
 import { routeNames } from '@site/constants';
 import { PageTitle, PageHeader } from '@site/components/text';
-import { parseFiltersFromSearchParams } from '@site/components/cards/CardListFilter';
+import { parseFiltersFromSearchParams } from '@site/components/cards/CardList';
 
 const designs = getAllCardDesigns();
 const filters = parseFiltersFromSearchParams(Astro.url.searchParams);
