+++ packages/site/src/pages/user/[username].astro
@@ -10,7 +10,7 @@ import { Anchor, Heading } from '@site/components/text';
 import UserLookingFor from '@site/components/user/UserLookingFor';
 import Card from '@site/components/cards/Card';
 import type { Props as HeaderProps } from '@site/components/Header.astro';
-import { parseFiltersFromSearchParams } from '@site/components/cards/CardListFilter';
+import { parseFiltersFromSearchParams } from '@site/components/cards/CardList';
 
 type Logo = HeaderProps['logo'];
